refactor(executor): split command.py and command.ts into a command package
