+++ python/mirage/workspace/executor/command/flags.py
@@ -1,0 +1,170 @@
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+from mirage.commands.spec import (CommandSpec, OperandKind, flag_kwarg_name,
+                                  parse_command, parse_to_kwargs)
+from mirage.commands.spec.usage import (invalid_argument_error,
+                                        missing_required_error,
+                                        missing_value_error,
+                                        unknown_option_error)
+from mirage.types import PathSpec
+from mirage.workspace.executor.command.types import ParsedCommand
+
+
+def synthesize_path_spec(value: str) -> PathSpec:
+    """A PathSpec for a path the classifier never saw.
+
+    Covers a relative value cwd-resolved by ``parse_command`` (e.g.
+    ``csplit -f part`` -> ``/data/part``) and a spec-classified PATH
+    operand the upstream classifier left as text. ``resource_path``
+    stays empty on purpose: the mount stamps the backend key on every
+    positional path and path-shaped flag value at execute time
+    (``Mount.execute_cmd``), so a parse-time stamp is dead weight —
+    proven by running the full suite with this field set to a
+    sentinel.
+
+    Args:
+        value (str): the resolved absolute virtual path.
+    """
+    return PathSpec(virtual=value,
+                    directory=value[:value.rfind("/") + 1] or "/",
+                    resource_path="",
+                    resolved=True)
+
+
+def parse_flags(
+    parts: list[str | PathSpec],
+    spec: CommandSpec | None,
+    cmd_name: str,
+    cwd: str,
+    str_flag_paths: bool = False,
+) -> ParsedCommand:
+    """Parse flags from classified parts, recovering PathSpec for PATH values.
+
+    Single-mount dispatch and cross-mount dispatch both parse through
+    here, so flags, texts, and parser warnings cannot drift between the
+    two paths (a cross-mount `grep --bogus` used to lose its warning).
+
+    Args:
+        parts (list[str | PathSpec]): expanded command words after the
+            command name; path-classified words arrive as PathSpec.
+        spec (CommandSpec | None): command spec, from the owning mount on
+            the single-mount path or the shared SPECS registry on the
+            cross-mount path; None falls back to type separation.
+        cmd_name (str): command name used in warnings.
+        cwd (str): current working directory for relative path resolution.
+        str_flag_paths (bool): keep PATH flag values as their resolved
+            virtual-path strings instead of PathSpec. Cross-mount
+            strategies read flags through FlagView, which type-checks
+            str, so they get the string view.
+
+    Returns:
+        ParsedCommand: positional paths, positional texts, parsed flag dict
+        (PATH flag values recovered to PathSpec, multiple PATH flags to
+        list[PathSpec]), and parser warnings (e.g. ignored unknown options).
+    """
+    # Build string argv and PathSpec lookup
+    argv = [
+        item.virtual if isinstance(item, PathSpec) else item for item in parts
+    ]
+    scope_map: dict[str, PathSpec] = {}
+    for item in parts:
+        if isinstance(item, PathSpec):
+            scope_map[item.virtual] = item
+            stripped = item.virtual.rstrip("/")
+            if stripped and stripped != item.virtual:
+                scope_map[stripped] = item
+
+    if spec is not None:
+        parsed = parse_command(spec, argv, cwd=cwd)
+        flag_kwargs = parse_to_kwargs(parsed)
+
+        # Recover PathSpec for PATH flag values; multiple PATH flags
+        # arrive as a list of resolved paths and become list[PathSpec].
+        # A relative PATH flag value cwd-resolved by parse_command (e.g.
+        # csplit -f part -> /data/part) is absent from scope_map, so build a
+        # PathSpec for it just like positional paths do, otherwise it never
+        # gets the mount prefix stripped.
+        repeat_path_keys = {
+            flag_kwarg_name(name)
+            for opt in spec.options
+            if opt.value_kind == OperandKind.PATH and opt.multiple
+            for name in (opt.short, opt.long) if name
+        }
+        single_path_keys = {
+            flag_kwarg_name(name)
+            for opt in spec.options
+            if opt.value_kind == OperandKind.PATH and not opt.multiple
+            for name in (opt.short, opt.long) if name
+        }
+        if not str_flag_paths:
+            for key, value in flag_kwargs.items():
+                if key in repeat_path_keys and isinstance(value, list):
+                    flag_kwargs[key] = [
+                        scope_map.get(part, synthesize_path_spec(part))
+                        for part in value
+                    ]
+                elif key in single_path_keys and isinstance(value, str):
+                    flag_kwargs[key] = scope_map.get(
+                        value, synthesize_path_spec(value))
+                elif isinstance(value, str) and value in scope_map:
+                    flag_kwargs[key] = scope_map[value]
+
+        # Classify positional args
+        paths: list[PathSpec] = []
+        texts: list[str] = []
+        for value, kind in parsed.args:
+            if kind == OperandKind.PATH:
+                scope = scope_map.get(value)
+                if scope is None:
+                    scope = synthesize_path_spec(value)
+                paths.append(scope)
+            else:
+                texts.append(value)
+        return ParsedCommand(paths, texts, flag_kwargs, parsed.warnings,
+                             parsed.invalid_options,
+                             parsed.needs_value_options,
+                             parsed.invalid_value_options,
+                             parsed.missing_required_options)
+
+    # No spec: separate by type
+    paths = [item for item in parts if isinstance(item, PathSpec)]
+    texts = [item for item in parts if not isinstance(item, PathSpec)]
+    return ParsedCommand(paths, texts, {}, [], [], [], [], [])
+
+
+def option_error(cmd_name: str,
+                 parsed: ParsedCommand) -> tuple[bytes, int] | None:
+    """GNU-shaped refusal for option errors the parser reported.
+
+    find is exempt: its expression tokens are validated by
+    parse_find_expression, which raises the GNU predicate error itself.
+
+    Args:
+        cmd_name (str): command name for message shape and exit code.
+        parsed (ParsedCommand): parse result carrying the reports.
+    """
+    if cmd_name == "find":
+        return None
+    if parsed.invalid_options:
+        return unknown_option_error(cmd_name, parsed.invalid_options[0])
+    if parsed.needs_value_options:
+        return missing_value_error(cmd_name, parsed.needs_value_options[0])
+    if parsed.invalid_value_options:
+        option, value, choices = parsed.invalid_value_options[0]
+        return invalid_argument_error(cmd_name, option, value, choices)
+    if parsed.missing_required_options:
+        return missing_required_error(cmd_name,
+                                      parsed.missing_required_options[0])
+    return None
