+++ typescript/packages/core/src/workspace/executor/command/types.ts
@@ -1,0 +1,19 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import type { ByteSource, IOResult } from '../../../io/types.ts'
+import type { ExecutionNode } from '../../types.ts'
+
+export type Result = [ByteSource | null, IOResult, ExecutionNode]
+export type Flags = Record<string, string | boolean | number | string[]>
