+++ typescript/packages/core/src/workspace/node/command_dispatch.ts
@@ -157,7 +157,6 @@ export async function executeCommand(
   callStack: CallStack | null,
   jobTable: JobTable | null,
   ensureOpen?: (resource: Resource) => Promise<void>,
-  unmount?: (prefix: string) => Promise<void>,
   runtimeBindings?: Record<string, Runtime>,
   routingDecision?: PolicyDecision,
   signal?: AbortSignal,
@@ -218,7 +217,6 @@ export async function executeCommand(
       callStack,
       jobTable,
       ensureOpen,
-      unmount,
       runtimeBindings,
       routingDecision,
       signal,
@@ -254,7 +252,6 @@ async function runCommandBody(
   callStack: CallStack | null,
   jobTable: JobTable | null,
   ensureOpen?: (resource: Resource) => Promise<void>,
-  unmount?: (prefix: string) => Promise<void>,
   runtimeBindings?: Record<string, Runtime>,
   routingDecision?: PolicyDecision,
   signalIn?: AbortSignal,
@@ -338,7 +335,6 @@ async function runCommandBody(
       callStack,
       jobTable,
       ensureOpen,
-      unmount,
       runtimeBindings,
       routingDecision,
       signal,
@@ -387,7 +383,6 @@ async function runArgv(
   callStack: CallStack | null,
   jobTable: JobTable | null,
   ensureOpen?: (resource: Resource) => Promise<void>,
-  unmount?: (prefix: string) => Promise<void>,
   runtimeBindings?: Record<string, Runtime>,
   routingDecision?: PolicyDecision,
   signal?: AbortSignal,
@@ -694,7 +689,6 @@ async function runArgv(
     callStack,
     jobTable,
     ensureOpen,
-    unmount,
     runtimeBindings,
     namespace,
     routingDecision,
